Added R (roxygen2), demo and description files.
The license file apparently is not needed for standard licenses, hence it was deleted. The SHA entry in the description file was removed as the SHA is apparently renewed when doing an amended commit. This apparently makes it impossible to get the current SHA in the DESCRIPTION file. This is the first complete working version of the package.